backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
numpy>=1.26.0
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10
